base-requirements.txt
Django==1.5
mimeparse==0.1.3
lxml==3.0.1
pyyaml==3.10
uuid==1.30
django-tastypie==0.9.15
south==0.7.6
celery==3.0.19
django-celery==3.0.11
celery-with-redis==3.0
gunicorn==0.17.2
dogstatsd-python==0.2
boto==2.6.0
django-pipeline==1.3.6
django-guardian==1.1
django-request-provider==1.0.2
django-haystack==1.2.7
Whoosh==2.4.1
nose==1.2.1
django-nose==1.1
path.py==3.1
django-allauth==0.10.1
python-memcached==1.48
django-ses==0.4.1
django-analytical==0.15.0
mock==1.0.1

